Time: 12 ms (39.26%) | Memory: 10.8 MB (59.81%) - LeetSync

+++ 2819-remove-trailing-zeros-from-a-string/remove-trailing-zeros-from-a-string.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string removeTrailingZeros(string num) {
-        size_t n=num.size();
+        int n=num.size();
         while (num.back() == '0') {
             num.replace(n-1,1,"");
             n--;
